requirements.txt
Flask==0.9
Jinja2==2.6
Werkzeug==0.8.3
argparse==1.2.1
distribute==0.6.24
gunicorn==0.17.2
py-bcrypt==0.2
redis==2.7.2
wsgiref==0.1.2
